refactor test

+++ src/main/java/com/ginsan/shape/Rectangle.java
@@ -1,18 +1,18 @@
 package com.ginsan.shape;
 
 public class Rectangle {
-    private int height ;
-    private int width;
+    private double height ;
+    private double width;
 
-    public Rectangle(int length, int width) {
-        this.height = Integer.max(length,0);
-        this.width = Integer.max(width,0);
+    public Rectangle(double length, double width) {
+        this.height = Double.max(length,0);
+        this.width = Double.max(width,0);
     }
 
-    public int getPerimeter(){
+    public double getPerimeter(){
         return (height+width)*2;
     }
-    public int getArea(){
+    public double getArea(){
         return height*width;
     }
